% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{dshift}
\alias{dshift}
\title{Take first-difference transformation of a series}
\usage{
dshift(x)
}
\arguments{
\item{x}{the series to be first-differenced}
}
\value{
the first-differenced series
}
\description{
Take first-difference transformation of a series
}
\details{
\code{dshift} assumes that the series are ordered, that there is no missing
data, and that the time intervals are even.
}
\examples{
dshift(rnorm(20))
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{utilities}
