% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{dynamac-deprecated}
\alias{dynamac-deprecated}
\title{Deprecated functions in package \pkg{dynamac}}
\description{
The functions listed below are deprecated and will be defunct
in the near future. When possible, functions with similar functionality are mentioned.
Help pages for deprecated functions are available at \code{help-("deprecated")}.
}
\keyword{internal}
