% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{dynardl.auto.correlated}
\alias{dynardl.auto.correlated}
\title{Perform post-estimation tests for white-noise residuals}
\usage{
dynardl.auto.correlated(x, bg.type = "Chisq", digits = 3, order = NULL,
  object.out = FALSE)
}
\arguments{
\item{x}{a \code{dynardl} model}

\item{bg.type}{a character string for the type of Breusch-Godfrey test to
run. The default is \code{Chisq}: the Chisq test statistic. The other option
is \code{F}: the F-test statistic.}

\item{digits}{number of digits to round when showing output (rarely used).
We recommend three.}

\item{order}{test for up to p'th order of AR(p) autocorrelation when
executing the Breusch-Godfrey test}

\item{object.out}{if \code{TRUE}, and \code{dynardl.auto.correlated} is
assigned to an object, the AIC, BIC, and results will be stored for the
user's convenience.}
}
\value{
The results of autocorrelation tests.
}
\description{
Perform post-estimation tests for white-noise residuals
}
\details{
\code{dynardl.auto.correlated} performs post-estimation diagnostics designed
to help users with model specification (as an aside, this is also why
\code{dynardl} has a \code{simulate = FALSE} argument; users can ensure the
model is white noise residuals before estimating a potentially
time-intensive simulation). These include the Breusch-Godfrey and
Durbin-Watson tests for autocorrelation, Shapiro-Wilk test for normality,
and various information criteria. The output reminds the user of the null
hypotheses for the autocorrelation tests.
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{post-estimation}
\keyword{utilities}
