% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\docType{data}
\name{france.data}
\alias{france.data}
\title{Data on French Energy Consumption and GDP}
\format{A data frame with 53 rows and 4 variables:
\describe{
  \item{country}{Country}
  \item{year}{Year}
  \item{lnGDP_cons2010USD}{ln(GDP), constant 2010 US dollars}
  \item{lnenergy}{ln(energy consumption), mill. tons oil equiv.}
}}
\usage{
data(france.data)
}
\description{
Data on GDP are from World Bank World Development Indicators. Data
on energy consumption are from the PB Statistical Review of World
Energy (June 2018).
}
\keyword{datasets}
