% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\docType{data}
\name{ineq}
\alias{ineq}
\title{Data on public concern about economic inequality}
\format{A data frame with 49 rows and 9 variables:
\describe{
  \item{year}{Year}
  \item{mood}{Public mood liberalism}
  \item{urate}{Unemployment rate}
  \item{concern}{Concern about economic inequality}
  \item{demcontrol}{Democratic control of congress}
  \item{incshare10}{Proportion of income of top 10 percent}
  \item{csentiment}{Consumer sentiment}
  \item{incshare01}{Proportion of income of top 1 percent}
}}
\source{
\url{http://dx.doi.org/10.7910/DVN/UYUU9G}
}
\usage{
data(ineq)
}
\description{
A dataset from: Wright, Graham. 2017. "The political
implications of American concerns about economic inequality."
Political Behavior 40(2): 321-346.
}
\keyword{datasets}
