% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{ldshift}
\alias{ldshift}
\title{Take lagged first-difference transformation of a series}
\usage{
ldshift(x, l)
}
\arguments{
\item{x}{the series to be first-differenced and then lagged}

\item{l}{the number of lags}
}
\value{
the lagged first-differenced series
}
\description{
Take lagged first-difference transformation of a series
}
\details{
\code{ldshift} assumes that the series are ordered, that there is no missing
data, and that the time intervals are even.
}
\examples{
ldshift(rnorm(20), l = 1)
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{utilities}
