% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{pssbounds}
\alias{pssbounds}
\title{Perform Pesaran, Shin and Smith (2001) cointegration test}
\usage{
pssbounds(data = list(), obs = NULL, fstat = NULL, tstat = NULL,
  case = NULL, k = NULL, digits = 3, object.out = FALSE)
}
\arguments{
\item{data}{data to be passed through from dynardl estimation}

\item{obs}{number of observations}

\item{fstat}{F-statistic of the joint test that \code{l.y = l.x1 + l.x2 +...+l.xk = 0}}

\item{tstat}{t-static of the one-sided}

\item{case}{specify certain restrictions on the constant and trend terms, since critical values differ by case. Case I: no intercept or trend, Case II: restricted intercept, no trend, Case III: unrestricted intercept with no trend, Case IV: unrestricted intercept and restricted trend, Case V: unrestricted intercept and trend. Case III is most frequently specified}

\item{k}{number of regressors appearing in levels in the estimated  model}

\item{digits}{number of digits to present in output (rarely used)}

\item{object.out}{output results (rarely used)}
}
\description{
Perform Pesaran, Shin and Smith (2001) cointegration test
}
\details{
pssbounds performs post-estimation cointegration testing using the bounds testing procedure from Pesaran, Shin, and Smith (2001). Since test statistics vary based on the number of \code{k} regressors, length of the series, these are required to specify, in addition to F- and t-statistics.
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{cointegration}
