% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\docType{data}
\name{supreme.sup}
\alias{supreme.sup}
\title{Data on US Supreme Court Approval}
\format{A data frame with 42 rows and 9 variables:
\describe{
  \item{dcalc}{Supreme Court support}
  \item{l_dcalc}{Lagged Supreme Court spport}
  \item{iddiv}{Ideological divergence}
  \item{mooddev}{Mean deviation of Mood}
  \item{dirdev}{Mean deviation of percent liberal decisions}
  \item{sg}{Rulings against Solicitor General's amicus briefs}
  \item{laws}{Laws declared unconstitutional}
  \item{presapp}{Approval of president}
  \item{congapp}{Approval of Congress}
}}
\source{
\url{http://dx.doi.org/10.2307/2669280}
}
\usage{
data(supreme.sup)
}
\description{
A dataset from: Durr, Robert H., Andrew D. Martin, and Christina Wolbrecht. 2000. "Ideological divergence and public support for the Supreme Court." American Journal of Policial Science 44(4): 768-776.
}
\keyword{datasets}
