% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{dshift}
\alias{dshift}
\title{Take first difference of a series.}
\usage{
dshift(x)
}
\arguments{
\item{x}{a series to be differenced}
}
\value{
the differenced series
}
\description{
Take first difference of a series.
}
\details{
\code{dshift} assumes that the series are ordered, that there is no missing data, and that the time intervals are even.
}
\examples{
x.var <- seq(0, 50, 5)
d.x.var <- dshift(x.var)
head(x.var)
head(d.x.var)
}
\author{
Soren Jordan and Andrew Q. Philips
}
\keyword{utilities}
