% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamac.R
\name{dynamac-deprecated}
\alias{dynamac-deprecated}
\alias{area.simulation.plot}
\alias{spike.simulation.plot}
\title{Deprecated functions in package \pkg{dynamac}}
\usage{
area.simulation.plot(x, response = "levels", bw = FALSE)

spike.simulation.plot(x, response = "levels", bw = FALSE)
}
\arguments{
\item{x}{a dynardl model with a simulation to be plotted}

\item{response}{whether the plot of the response should be shown in levels of the dependent variable (\code{levels}) 
or in changes from the mean of the dependent variable (\code{mean.changes}). The default is \code{levels}}

\item{bw}{should the colors be in black and white (for publication)? The default is \code{FALSE}}

\item{x}{a dynardl model with a simulation to be plotted}

\item{response}{whether the plot of the response should be shown in levels of the dependent variable (\code{levels}) 
or in changes from the mean of the dependent variable (\code{mean.changes}). The default is \code{levels}}

\item{bw}{should the colors be in black and white (for publication)? The default is \code{FALSE}}
}
\value{
an area plot

a spike plot
}
\description{
The functions listed below are deprecated and will be defunct
in the near future. When possible, functions with similar functionality are mentioned.
Help pages for deprecated functions are available at \code{help-("deprecated")}.
}
\section{\code{area.simulation.plot}}{

For \code{area.simulation.plot}, use \code{\link{dynardl.simulation.plot}}.
}

\section{\code{spike.simulation.plot}}{

For \code{spike.simulation.plot}, use \code{\link{dynardl.simulation.plot}}.
}

\seealso{
\code{link{dynamac-deprecated}}

\code{link{dynamac-deprecated}}
}
\author{
Soren Jordan and Andrew Q. Philips

Soren Jordan and Andrew Q. Philips
}
\keyword{internal}
\keyword{utilities}
